+++ coauthor/agent/agent_base.py
@@ -31,7 +31,7 @@ class BaseReflectChainAgent(ABC):
     def __init__(self, model_handler: Any, agent_config: AgentConfig, agent_settings: AgentSettings, agent_prompts: AgentPrompts, agent_path: str):
         """Initialize with model handler, agent config, settings/prompts, and agent path"""
         self.model_handler = model_handler
-        self.agent_config = agent_config 
+        self.agent_config = agent_config
         self.agent_settings = agent_settings
         self.agent_prompts = agent_prompts
         self.agent_path = agent_path
@@ -316,7 +316,7 @@ class BaseReflectChainAgent(ABC):
         system_prompt = render_prompt(self.agent_prompts.system_prompt, user_vars)
         user_request = render_prompt(self.agent_prompts.user_request, user_vars)
         user_prefix = render_prompt(self.agent_prompts.user_prefix, user_vars)
-    
+
         # TODO: Consider making tex_count_stats and first_k_tex_document part of a agent tool state object
         if tex_count_stats:
             user_prefix = f"{tex_count_stats}{user_prefix}"
@@ -331,7 +331,7 @@ class BaseReflectChainAgent(ABC):
 
         prefill = self.agent_settings.prefills[round] if round < len(self.agent_settings.prefills) else self.agent_settings.prefills[0]
         accumulated_output = prefill if prefill else ""
-    
+
         accumulated_output, end_turn, messages = self.model_handler.initialize_output_and_prefill(
             output_file,
             self.agent_config,
