+++ coauthor/agent/model_registry.py
@@ -16,10 +16,7 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         context_window=200000,
         input_price=15.0,
         output_price=75.0,
-        capabilities=ModelCapabilities(
-            supports_prompt_caching=True,
-            supports_assistant_prefill=True
-        )
+        capabilities=ModelCapabilities(supports_prompt_caching=True, supports_assistant_prefill=True),
     ),
     "sonnet++": AnthropicModelConfig(
         name="sonnet++",
@@ -29,11 +26,8 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         input_price=3.0,
         output_price=15.0,
         capabilities=ModelCapabilities(
-            supports_prompt_caching=True,
-            supports_native_pdf=True,
-            supports_assistant_prefill=True,
-            likes_to_ask_for_confirmation=True
-        )
+            supports_prompt_caching=True, supports_native_pdf=True, supports_assistant_prefill=True, likes_to_ask_for_confirmation=True
+        ),
     ),
     "sonnet+": AnthropicModelConfig(
         name="sonnet+",
@@ -42,10 +36,7 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         context_window=200000,
         input_price=3.0,
         output_price=15.0,
-        capabilities=ModelCapabilities(
-            supports_prompt_caching=True,
-            supports_assistant_prefill=True
-        )
+        capabilities=ModelCapabilities(supports_prompt_caching=True, supports_assistant_prefill=True),
     ),
     "sonnet": AnthropicModelConfig(
         name="sonnet",
@@ -54,10 +45,7 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         context_window=200000,
         input_price=3.0,
         output_price=15.0,
-        capabilities=ModelCapabilities(
-            supports_prompt_caching=False,
-            supports_assistant_prefill=True
-        )
+        capabilities=ModelCapabilities(supports_prompt_caching=False, supports_assistant_prefill=True),
     ),
     "haiku+": AnthropicModelConfig(
         name="haiku+",
@@ -67,11 +55,8 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         input_price=1.0,
         output_price=5.0,
         capabilities=ModelCapabilities(
-            supports_prompt_caching=True,
-            supports_vision=False,
-            supports_assistant_prefill=True,
-            likes_to_ask_for_confirmation=True
-        )
+            supports_prompt_caching=True, supports_vision=False, supports_assistant_prefill=True, likes_to_ask_for_confirmation=True
+        ),
     ),
     "haiku": AnthropicModelConfig(
         name="haiku",
@@ -80,10 +65,7 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         context_window=200000,
         input_price=0.25,
         output_price=1.25,
-        capabilities=ModelCapabilities(
-            supports_prompt_caching=True,
-            supports_assistant_prefill=True
-        )
+        capabilities=ModelCapabilities(supports_prompt_caching=True, supports_assistant_prefill=True),
     ),
     # OpenAI models
     "gpto1": OpenAIModelConfig(
@@ -93,11 +75,7 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         context_window=128000,
         input_price=15.0,
         output_price=60.0,
-        capabilities=ModelCapabilities(
-            supports_auto_prompt_caching=True,
-            supports_vision=False,
-            supports_reasoning=True
-        )
+        capabilities=ModelCapabilities(supports_auto_prompt_caching=True, supports_vision=False, supports_reasoning=True),
     ),
     "gpto1-": OpenAIModelConfig(
         name="gpto1-",
@@ -106,11 +84,7 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         context_window=128000,
         input_price=3.0,
         output_price=12.0,
-        capabilities=ModelCapabilities(
-            supports_prompt_caching=True,
-            supports_vision=False,
-            supports_reasoning=True
-        )
+        capabilities=ModelCapabilities(supports_prompt_caching=True, supports_vision=False, supports_reasoning=True),
     ),
     "gpt4o": OpenAIModelConfig(
         name="gpt4o",
@@ -121,10 +95,7 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         context_window=128000,
         input_price=2.5,
         output_price=10.0,
-        capabilities=ModelCapabilities(
-            supports_auto_prompt_caching=True,
-            supports_predictive_output=True
-        )
+        capabilities=ModelCapabilities(supports_auto_prompt_caching=True, supports_predictive_output=True),
     ),
     "gpt4t": OpenAIModelConfig(
         name="gpt4t",
@@ -133,10 +104,7 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         context_window=128000,
         input_price=10.0,
         output_price=30.0,
-        capabilities=ModelCapabilities(
-            supports_auto_prompt_caching=True,
-            supports_reasoning=True
-        )
+        capabilities=ModelCapabilities(supports_auto_prompt_caching=True, supports_reasoning=True),
     ),
     "gpt4o-": OpenAIModelConfig(
         name="gpt4o-",
@@ -145,10 +113,7 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         context_window=128000,
         input_price=0.15,
         output_price=0.6,
-        capabilities=ModelCapabilities(
-            supports_auto_prompt_caching=True,
-            supports_predictive_output=True
-        )
+        capabilities=ModelCapabilities(supports_auto_prompt_caching=True, supports_predictive_output=True),
     ),
     "gpt4ol": OpenAIModelConfig(
         name="gpt4ol",
@@ -157,10 +122,7 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         context_window=128000,
         input_price=5.0,
         output_price=15.0,
-        capabilities=ModelCapabilities(
-            supports_auto_prompt_caching=True,
-            supports_reasoning=True
-        )
+        capabilities=ModelCapabilities(supports_auto_prompt_caching=True, supports_reasoning=True),
     ),
     "geminiexp": OpenAICompatibleModelConfig(
         name="geminiexp",
@@ -170,10 +132,7 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         context_window=2097152,
         input_price=1.25,
         output_price=5.0,
-        capabilities=ModelCapabilities(
-            supports_prompt_caching=True,
-            supports_assistant_prefill=True
-        )
+        capabilities=ModelCapabilities(supports_prompt_caching=True, supports_assistant_prefill=True),
     ),
     # Google Gemini models
     "gemini2f": OpenAICompatibleModelConfig(
@@ -184,10 +143,7 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         context_window=1048576,
         input_price=0.075,
         output_price=0.3,
-        capabilities=ModelCapabilities(
-            supports_prompt_caching=False,
-            supports_native_pdf=True
-        )
+        capabilities=ModelCapabilities(supports_prompt_caching=False, supports_native_pdf=True),
     ),
     "gemini1p+": OpenAICompatibleModelConfig(
         name="gemini1p+",
@@ -196,10 +152,7 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         max_output_tokens=8192,
         input_price=1.25,
         output_price=5.0,
-        capabilities=ModelCapabilities(
-            supports_prompt_caching=True,
-            supports_assistant_prefill=True
-        )
+        capabilities=ModelCapabilities(supports_prompt_caching=True, supports_assistant_prefill=True),
     ),
     "gemini1f+": OpenAICompatibleModelConfig(
         name="gemini1f+",
@@ -209,10 +162,7 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         context_window=1048576,
         input_price=0.075,
         output_price=0.3,
-        capabilities=ModelCapabilities(
-            supports_prompt_caching=True,
-            supports_native_pdf=True
-        )
+        capabilities=ModelCapabilities(supports_prompt_caching=True, supports_native_pdf=True),
     ),
     # OpenRouter models
     "gpt4oOR": OpenAICompatibleModelConfig(
@@ -223,10 +173,7 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         context_window=128000,
         input_price=6.0,
         output_price=18.0,
-        capabilities=ModelCapabilities(
-            supports_prompt_caching=True,
-            supports_reasoning=True
-        )
+        capabilities=ModelCapabilities(supports_prompt_caching=True, supports_reasoning=True),
     ),
     "gemini1p+OR": OpenAICompatibleModelConfig(
         name="gemini1p+OR",
@@ -236,10 +183,7 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         context_window=2097152,
         input_price=2.5,
         output_price=7.5,
-        capabilities=ModelCapabilities(
-            supports_prompt_caching=True,
-            supports_assistant_prefill=True
-        )
+        capabilities=ModelCapabilities(supports_prompt_caching=True, supports_assistant_prefill=True),
     ),
     "gemini1f+OR": OpenAICompatibleModelConfig(
         name="gemini1f+OR",
@@ -249,10 +193,7 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         context_window=1048576,
         input_price=0.075,
         output_price=0.3,
-        capabilities=ModelCapabilities(
-            supports_prompt_caching=True,
-            supports_native_pdf=True
-        )
+        capabilities=ModelCapabilities(supports_prompt_caching=True, supports_native_pdf=True),
     ),
     "llama3+OR": OpenAICompatibleModelConfig(
         name="llama3+OR",
@@ -262,10 +203,7 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         context_window=131072,
         input_price=3.0,
         output_price=3.0,
-        capabilities=ModelCapabilities(
-            supports_prompt_caching=True,
-            supports_reasoning=True
-        )
+        capabilities=ModelCapabilities(supports_prompt_caching=True, supports_reasoning=True),
     ),
     "qwq-32bOR": OpenAICompatibleModelConfig(
         name="qwq-32b",
@@ -275,10 +213,7 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         context_window=32768,
         input_price=0.15,
         output_price=0.6,
-        capabilities=ModelCapabilities(
-            supports_prompt_caching=True,
-            supports_reasoning=True
-        )
+        capabilities=ModelCapabilities(supports_prompt_caching=True, supports_reasoning=True),
     ),
     "sonnet++OR": OpenAICompatibleModelConfig(
         name="sonnet++OR",
@@ -288,9 +223,6 @@ MODEL_HANDLERS: Dict[str, ModelConfig] = {
         context_window=200000,
         input_price=3.0,
         output_price=15.0,
-        capabilities=ModelCapabilities(
-            supports_prompt_caching=True,
-            supports_reasoning=True
-        )
+        capabilities=ModelCapabilities(supports_prompt_caching=True, supports_reasoning=True),
     ),
 }
